(2.6) dcache-webadmin: change Jetty setting so .war is not unpacked
See http://rt.dcache.org/Ticket/Display.html?id=8489.

When upgrading from 2.6 to 2.10, there is a potential
for the webadmin directory timestamps to be more recent
than the timestamp of the /usr/share/dcache/classes/webapps/webadmin.war
file, which then causes Jetty to ignore the new war
and use the old directory. This subsequently provokes
jndi issues because the new httpd.batch file does not
pass through properties that the webadmin spring file
expects to find in the naming and directory interface.

The simplest solution is to turn off unpacking.

While the specific issue reported in the ticket does not
occur in 2.6, the conversion to non-extraction was requested
for this branch as well, to prevent further possible
timestamp issues on the update of the .war.

Note that https://rb.dcache.org/r/7436/ does this for master.

Target: 2.6
Acked-by: Gerd
Require-book: no
Require-notes: yes

RELEASE NOTES:  The webadmin.war found in /usr/share/dcache/classes/webapps
is no longer unpacked (by default, /var/lib/dcache/httpd).  This
avoids potential conflicts between versions when update installs are run.