httpd: Drop info pages from webadmin
The info pages exist because originally webadmin was developed as a complete
replacement for httpd, meaning everything httpd could do, webadmin had to do
too, even if that functionallity wasn't really a part of the admin tool.

The info for info-provider is just as easy to expose through the classic
info link - info-provider was actually never updated to use webadmin.

Even the UI in webadmin for accessing the info xml is inferior to just
load the old page in the browser.

This patch removes both the Info and InfoXml pages from webadmin.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: https://rb.dcache.org/r/7536/