pool: Simplify synchronization in nearline storage handler
To allow parallellizaation in nearline storage request creation,
2.11 introduced the use of a Guava Striped object to lock a part
of the request key space. This was necessary as we didn't want
to rely on Java 8 in dCache 2.11. With master now using Java 8,
the code can be simplified by using ConcurrentHashMap's ability
to compute the value to be inserted on demand when not already
present. The map imeplementation locks hash table buckets, which
is potentially more fine-grained as the Striped object used.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7559/