loginbroker: IPv6 enable and other improvements
This patch makes several improvements how doors publish their network
interfaces. It doeesn't actually touch LoginBroker itself; rather it updates
which information doors publish and how consumers use it.

First of all, the patch makes interface discovery dynamic. If a door is
configured to use a wild card address (ie it listens on all interfaces and all
addresses), the LoginBrokerHandler will now discover when interfaces go up and
down. This means that we cannot use NetworkUtils, as that class statically
discovers local addresses on startup and never changes the list.

The second change is that rather than just publishing a list of strings, which
may be host names or addresses, the doors now publish a list of resolved
InetAddress objects. By resolved I mean that the object contains both the
address and a cached host name. Care is taken that any FQDN an admin may have
specified in the configuration survives to the published information.

Except when the custom-get-host-by-address feature is enabled (can anybody
remember what that was for?), the SRM no longer performs a reverse-lookup
on the door's address. This means that it is entirely the name configuration
of the door or the name specified in the door by the admin that controls
the host name that is included in TURLs.

The SRM no longer limits itself to just choosing the first in the list of
addresses of the door. Instead it chooses an address that has at least a big a
scope as the client's address. That is, if a client connects from a global
address, the SRM will produce a TURL with a host resolving to a global address
too. If a client connects from a site local address, the SRM may use a name
resolving to a site local address. In fact, the SRM will prefer the most
limited scope that is at least as big as the client's scope. That is, if a
client connects from a site address, the SRM will choose the site address of
the door, if one exists. This will break down for an installation covering
multiple disconnected private networks (eg NDGF), but on other sites this will
mean that a local client gets a local TURL and an external client a global
TURL. Since most doors too choose a pool interface based on whehter the client
connected to the internal or external interface of the door, this change nicely
extends the concept to the SRM.

Similar logic is used for chosing the IP protocol family (ie IPv4 or IPv6). If
the SRM client connects over IPv4, the SRM will try to choose an IPv4 address.
If the client connects over IPv6, the SRM will try to choose an IPv6 address.
This is only a soft restriction: If no matching door is available for the scope
and protocols needed, the SRM falls back to the other family. Thus an IPv6
client may get an IPv4 TURL and it is then up to the client to fail the request
if it isn't dual-stacked.

It is important to note that TURLs still contain hostnames (FQDN actually).
Although using IP addresses would be more precise (you never know if the name
resolves to the IP you expected or whether the client chooses the correct
protocol family if the name resolves to both IPv4 and IPv6 addresses), the host
certificates used for GSI and SSL based protocols likely only contains host
names. Thus we must return a TURL with a host that matches the certificate used
by the door.

A future improvement could be to actively use the hostname (both primary and
alt names) from the host certificate when publishing interfaces to LoginBroker.

A third change is that the interfaces published to LoginBroker are no longer
ordered. It is the consumer's responsibility to order according to its
preferences. For the info and httpd services, they attempt to use an external
interface. For the SRM, the logic described above is used. In fact, the info
service ought to be updated to not define an order at all. Instead the address
scope should be included in the interfaces published and the info provider
should only publish doors with an external interface.

LoginBrokerHandler will not publish the loopback address, except when
explicitly chosen by the admin as part of the *.net.listen property.

The *.net.listen properties have been modified to allow a hostname (FQDN) to be
specified rather than just an address. This name is used when publishing to
LoginBroker and thus also controls the name used by the info-provider and the
SRM. The info service no longer performs any reverse lookups on a door's
address.

For convenience, FsPath was moved to commons and LoginBrokerInfo was cleaned
up and made immutable.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7571/