scripts: fix and update wait-for-cells
The wait-for-cells script uses the info service to determine whether
the supplied list of cells is currently running.  If not, the script
will wait a defined period for the cells to appear before terminating
with an error.

In its current form, the script is broken.  This patch fixes the script,
drops the '.sh' prefix, and moves it into the sbin directory.

Some minor improvements are also made

Target: master
Patch: https://rb.dcache.org/r/7574/
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan
Requires-notes: yes
Requires-book: yes