loginmanager: Fix leak caused by absent child limit
A regression in 2.8 causes the door to not track child removal if the
connection limit is set to -1. This causes the door to eventually run out of
memory.

Rather than just restoring the pre-2.8 logic, the patch causes the login
manager to track children regardless of the limit. Even when the child count
isn't needed and even when the list of children isn't published to a login
broker, there are other features in the login manager that rely on it (such as
keep alive).

Target: trunk
Request: 2.11
Request: 2.10
Request: 2.9
Request: 2.8
Require-notes: yes
Require-book: no
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8570
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7585