dcap: Fix regression in published protocol family
Prior to 2.7, gsidcap was published as gsidcap and the SRM used this
as the protocol name when negotiating the transfer protocol. 2.7
always publishes 'dcap', which is a regression compared to 2.6.

This patch resolves the problem.

Target: trunk
Request: 2.11
Request: 2.10
Request: 2.9
Request: 2.8
Request: 2.7
Require-notes: yes
Require-book: no
Reported-by: Yvan Calas <yvan.calas@cc.in2p3.fr>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7587/