srm: allow sites to configure additional TExtraInfo for srmPing responses
An SRM server's response to a ping request allows the server to send
an arbitrary list of key-value pairs back to the client.  Currently,
dCache SRM is hard-coded to provide two such key-value pairs: the
implementation name ('dCache') and the dCache version.

This patch allows an admin to configure any number of additional
key-value pairs. This is achieved by the admin configuring dCache
properties that have a specific prefix.  This prefix is stripped and
the resulting key-value pairs are added to the SRM ping response.

The motivation is that we want to allow dCache to advertise
admin-configurable hints to SRM clients.  The most immediate one being
to allow dCache to suggest how many concurrent bring-online requests a
particular dCache instance can sustain.

As a consequence of this design, the patch introduces generic support
for prefix-based configuration.  This may be used in other places; for
example, when allowing admins to configure third-party libraries
(e.g., Hikari).

Known limitations of this patch:

  a. the "prefix" annotation (introduced to suppress info messages
     about non-system properties) and prefix used by
     ConfigurationMapFactoryBean are only "conincidentally" the same:
     any mismatch will not be detected.

  b. the prefix checking in ConfigurationProperties has terrible
     performance as the number of prefixes increases much beyond 1.
     This should be revisited if the prefix idea is reused elsewhere.

  c. the patch deliberately avoids using Java 8 constructs.  This is
     because we may want to back-port this to supported branches.

Target: master
Requires-book: yes
Requires-notes: yes
Patch: https://rb.dcache.org/r/7591/
Acked-by: Gerd Behrmann
Acked-by: Dmitry Litvintsev