ftp: Clean up address handling
Both the variable naming and how we compose various addresses is most
confusing. This patch tries to clean this up to make the logic easier to
follow.

There is also one behavioural change in how we determine the pool facing
interface for proxied data channels: If not set explicitly in the
configuration, before we defaulted to the control channel interface. Whether a
control channel is established on one or the other interface has absolutely no
relationship to which interface faces the pools, nor does establishing the
control channel on a different interface suddently make the pools available on
another interface. Hence this patch changes this to default to the default
interface of the host (as determined by resolving the host name).

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7594/