pool: Partially fix interface selection for FTP mover
The FTP mover tries to select an interface with the correct
protocol family and falls back to proxy mode if not possible.
Unfortunately the NetworkUtils utility class fails to return
null in case no suitable interface could be found.

On the other hand, the original protocol family less version
of the NetworkUtils method fails to handle null values.

This patch fixes the former to return null if the protocol
family could not be match, and fixes the latter by falling
back to ignoring the protocol family. The latter is likely
wrong, but it restore the previous functionality and resolves
the problem that existing callers of that method do not
handle a null return value.

There are plenty of issues with this piece of code. This
patch only sets out to fix the one.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.11
Request: 2.10
Request: 2.9
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7597/