Switch to using GlobusPrincipal from JGlobus 2
When we switched to JGlobus 2, we had to keep the old GlobusPrincipal
to maintain backwards compatibility with old pools. The pools we
need to maintain compatibility with now all use JGlobus 2, so we
can safely switch to the new verion.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7603/