webdav: Fix reported content length for partial GETs
When a client requests a byte range beyond the end of the file,
Milton fails to take the file length into account. The consequence
is that we claim to provide a longer body than we do and that
the reported entity length in the Content-Range header is wrong.

This patch resolves the issue by fixing the range in our
custom response handler.

Resolves a compatibility issue with ARC.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.11
Request: 2.10
Request: 2.9
Request: 2.8
Request: 2.7
Request: 2.6
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7615/