pnfsmanager: Don't report file size on create unless it is known
When creating a name space entry, it's size is initially zero.
PnfsManager includes this size in the returned FileAttributes
even though the size isn't actually the size of the target file
and isn't known.

This breaks proxied chunked webdav upload as the code extracts
the expected size from the file attributes and uses that
when talking to the pool. The size not being the actual size,
the write to the pool fails.

This patch solves the problem by not letting Chimera report
a size unless the attribute is actually set. The patch needs
to update the doors too, as they erroneously assumed that a
size is always provided.

The patch isn't backportable in its current form. A separate
patch for stable branches will be posted.

Resolves a compatibility problem with ARC.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7616/