pool: Fix regression preventing configuration to be reloaded
Upon reloading the pool configuration an error is produced when an nearline
storage was already defined in the existing configuration. The problem is
that nearline storage setup isn't reset before reloading the configuration.

This patch fixes the problem by not applying configuration changes during setup
(and reload) directly to the affected nearline storages. Instead the changes
are collected and then applied and merged with the existing configuration at
the end. This approach prevents that existing nearline stoarge requests have to
be cancelled in the process.

The HsmSet class would benefit from a split into a configuration class and a
user interface, but in the interest of making this back-portable, this patch
updates the existing class.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.11
Request: 2.10
Request: 2.9
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8501
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7622/