webdav,srm: provide useful log messages for SSL handshake failures.
Java SSL and GSS exceptions have a nasty habit of wrapping
an exception with useful information without including that
information in the wrapping exception's message.  Instead,
a generic phrase is returned, such as "General SSLEngine
problem".

The jGlobus base exception contains a work-around for this
problem by walking the exception chain to find the first
"interesting" exception.  Simple heuristics are used to
determine whether an exception is interesting.

With the change from jGlobus engine to our custom engine, the
subsequent logging aspect lost this work-around.

This patch takes the work-around from jGlobus and adds it to
dCache as a new aspect that alters how the getMessage method
of an SSLException or a GSSException works.  This aspect is
also weaved for subclasses of these exceptions, such as
SSLHandshakeException.  All other exceptions are unaffected.

Target: master
Request: 2.11
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/7637/
Acked-by: Gerd Behrmann