bootloader: Don't fail on missing cell name
If a pool name isn't defined in the configuration, the dCache fails due
to failure to parse the shell oracle.

This patch addresses this problem. check-config doesn't report the error
as there is nothing in the definition of a service that requires it
to start a cell and thus we cannot complain about lack of a cell name.
The pool will however report the lack of a name during startup.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.11
Request: 2.10
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7641/