pool: Automatically calculate a gap if not specified in the configuration
We hardcode a default gap of 4GB. This works fine for production pools,
however during testing it is the source to a common problem when the
pool size is smaller than the gap.

This patch introduces automatic calculation of the gap for small pools
if not explicitly defined in the configuration.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7643/