Incorporate DCAP and NFS style writes into Transfer class
The Transfer class was originally modelled on the FTP code. It offers two
methods, createNameSpaceEntry and readNameSpaceEntry for respectively upload
and downloading a file.

Neither DCAP nor NFS work like this though: They create the name space entry
independently from the actual transfer. Thus the name space entry must be read
for both uploads and downloads.  The DCAP door doesn't use the Transfer class,
however the NFS door works around the issue by relying on internal knowledge of
the Transfer class (that it can call readNameSpaceEntry even when wanting to
uploading a file). This workaround broke when I wanted to update
readNameSpaceEntry and the internal behaviour changed.

This patch extends the Transfer class to allow for DCAP and NFS style uploads.
The readNameSpaceEntry is now parameterised to indicate whether the caller
would accept an upload. Thus the NFS door avoids calling what was documented to
trigger a download. The change also allows the Transfer class to verify the
appropriate permissions (not relevant for the NFS door as it disables the
permission check anyway).

The change allows the detection of downloads of incomplete files to be moved
into the readNameSpaceEntry method. This in turn prevents doors from attempting
to access the missing size information for file entries that do not have it (as
would have been the case in the FTP door when it processes parameters for
partial reads).

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7649/