webdav: don't log a stack-track when proxy transfer is interrupted
Although dCache behaves correctly for proxied uploads and downloads,
if a client interrupts such a transfer then a stack-track is logged.
This is because Milton wraps an IOException in a RuntimeException.

To counter this, the EofException is converted to a
BadRequestException for PUT requests, resulting in a 400 response if
the client half-closes the connection.  If the TCP connection is
fully closed then the client doesn't receive any response.

Closing a TCP connection during GET request, will be reported as an
internal error, but the client will not get this message as the
TCP connection is already closed.

Target: master
Request: 2.11
Request: 2.10
Request: 2.9
Request: 2.8
Request: 2.7
Request: 2.6
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/7653
Acked-by: Gerd Behrmann