pool: Make EntryState#DESTROYED deprecated
A pool entry has both a REMOVED state and a DESTROYED state. While
the file is open, a REMOVED file isn't actually deleted from the pool.
Once the file is closed, it is deleted and moved to the DESTROYED
state.

However, the DESTROYED state is odd in that it represents the state
in which the entry doesn't exist. As such, having an instance in
that state doesn't make much sense. This was apparent in some of
the code too, as there were often "just in case" checks for DESTROYED
that should never be reachable. The state was also not persisted to
disk.

This patch marks the state deprecated and removes any use of it. The
REMOVED state still exists for an open replica that has been deleted,
and once it is closed the replica is deleted. However there is no
longer the redundant DESTROY state.

The enum value isn't removed yet as EntryState is used in some
messages.

There is an apparent change in what happens when opening a REMOVED
entry: A unit test previously asserted that this was allowed. It
worked when using the mock repository, but the real repositories
didn't allow this operation. The unit test has now been removed.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7659/