pool: Precompute rep ls -s output
The rep ls -s command is used to provide statistics per storage class for
files on a pool. The current implementation would iterate over the complete
set of files, which is a fairly slow process for large pools. The statistics
module does this every night and on large pools it regularly times out.

This patch adds a state change listener that maintains the statistics. Doing
it during state changes is easy as the cache entry is alrady loaded at that
point.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7667/