transfermanager: convert to Universal Spring Cell
This patch is anticipation of a few more changes in
TransferManager to make it use the same injection pattern
for DataNucleus as the other services.

Note that the refactoring is minimal.  I have not attempted to
restructure or rename anything, but simply added and removed
what was necessary in order to make the service run under
the UniversalSpringCell.

These are optimizations, and need not be backported.

Another patch for the DataNucleus injection follows.

Target: master
Require-notes: no
Require-book: no
Acked-by: Gerd