cells: Fix message timeout in case of thread pool overflows
Asynchronous message callbacks may be scheduled on an Executor. This executor
may be bounded, which means it may throw RejectedExecutionException. If we
never call the callback, we risk that there is state in the services that
doesn't get cleaned. Invoking one of the callback methods is thus essential for
dCache.

In one case we already had a solution for this: When unable to inject the task
for calling the delivery of a reply, the request is reinserted into the map
tracking asynchronous requests. Eventually this request will time out and the
timeout callback is called instead. In all other cases the code would not
recover from a RejectedExecutionException and simply log a stack trace.

This patch extends the existing solution to all callback invocations: If the
callback is rejected, the request is reinserted into the map to allow the
timeout callback to be called later (note that the timeout callback itself may
be rejected, in which case the request is put back into the map too).

In several cases the code failed to call the callback on the correct executor.
This has been fixed too. The same goes for restoring the CDC when invoking the
timeout callback.

I did consider whether we could fall back to in-thread execution of the
callback, but since the callback may perform blocking operations, it would be
very dangerous to invoke the callback from outside the correct thread pool.

Since the callback map is unbounded, there is a risk that it could grow very
large. If callbacks kept failing, this patch would increase the risk of the map
growing aggressively. That is however a different problem that could happen in
the old code already - the proper solution for that would be to reject sending
new requests rather than to drop the callbacks.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.11
Request: 2.10
Request: 2.9
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8585
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7682/
(cherry picked from commit c07c2f6e9ea4fd01352016670c20e5399427599e)