webdav: ghost-out files that are currently being uploaded
Currently, when a file is currently being uploaded, the webdav door
will generate a directory listing much like any other file; however,
the links to view or download this file will not work.

With this patch, when the file has no size then no links are
generated and the size isn't listed.  The file's name and
last modified timestamp are both shown, but "ghosted out".

NB. If an uploading file's intended size is known (possible with
HTTP and any SRM-initiated upload) then the file is listed as
a completed file.  This is a trade-off while completely
accurate information is prohibatively expensive.

Target: master
Patch: https://rb.dcache.org/r/7689/
Acked-by: Gerd Behrmann
Requires-book: no
Requires-notes: no