info-provider: don't rely on interface order
Previously, the info service made sure that the interfaces were ordered
by introducing an ad-hoc index metric.  For each door, the
info-provider selected the first interface and published it.

There are two problems: this publishes interfaces with non-global scope
if the door is bound to a non-global interface, and if a door has
multiple global interfaces then only one would be published.

With this patch, that ordering is removed and info-provider publishes
all interfaces with 'global' scope.  Doors with multiple global-scope
interfaces are published multiple times (once for each global
interface) and doors without any global-scope interface are not
published.

Target: master
Requires-notes: no
Requires-book: no