pool: update NFS transfer service to validate inactive movers
The NFS protocol provides a mechanism to detect dead client. In dCache,
this functionality implemented by binding mover to a nfs session on a first
IO operation. When a pool detects expired session, the corresponding mover
is terminated.

While this mechanism works in most situations, there are at least two conner
cases which must be covered:
  a) client never connects to a pool and, as a result, mover is not bound to
    a nfs session. In such situations, we will observe NFS movers, with
    zero bytes transfered size, which never get accessed.

  b) due to network errors the client my loose the connection to the pool,
  but still maintain its state with a door.  In such situation, client will assume
  that mover still exists and will attempt to use. Nevertheless, the pool may already
  killed the corresponding mover, and reject IO request with NFSERR_BAD_STATEID .
  By receiving NFSERR_BAD_STATEID the client will issue a new OPEN requests, but
  won't ask for new mover. This will endup into an infinite pool, pool will not accept
  this transfers.

To cover this conditions, the nfs transfer service is update to change his behavior.
The mover still get associated with a nfs session, but not killed when session expires.
In case of session expires, mover get decoupled from a session, but stays active.
A special task periodically scans for movers without sessions as queries corresponding
door for a validity. A mover get killed only if door does not held a valid open-state.

The change introduces a new message, which is generic enough to be used by other protocols
as well.

Acked-by: Gerd Behrmann
Target: master
(cherry picked from commit feea8e9f17069ac8f3806e6293a7e961957f2d90)
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>