webdav: make progress markers more robust
Both the emitting of progress markers and checking whether the client
disconnected relies on querying the current status of the transfer.  If
this request triggers a CacheException then neither will happen.

With this patch, the progress marker is emitted whether or not
the query was successful.  If the TransferManager's query to the pool
was unsuccessful then that error message is updated to make it
clear the problem came from the pool.

Target: master
Request: 2.11
Request: 2.10
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/7681
Acked-by: Gerd Behrmann